+++ Poker/Models/OmahaHand.cs
@@ -8,7 +8,7 @@ namespace Poker.Models
   public class OmahaHand : BaseHand
   {
     public override string Name { get { return "Omaha"; } }
-    public override byte CardCount { get { return 4; } }
+    public override int CardCount { get { return 4; } }
     public OmahaHand() : base() { }
     public OmahaHand(ulong cardsMask) : base(cardsMask) { }
     public OmahaHand(string cards) : base(cards) { }
