+++ Poker/Models/OneCardHand.cs
@@ -8,7 +8,7 @@ namespace Poker.Models
   public class OneCardHand : BaseHand
   {
     public override string Name { get { return "One Card"; } }
-    public override byte CardCount { get { return 1; } }
+    public override int CardCount { get { return 1; } }
     public OneCardHand() : base() { }
     public OneCardHand(ulong cardsMask) : base(cardsMask) { }
     public OneCardHand(string cards) : base(cards) { }
